Refactored Most physics body creation. placed it under a IWorld interface to centralize it all.
IWorld interface now returns gamemode and gamestate this is so that controllers and actors can request an implementation of IWorld to be able to do operations on the physics world without having a dependancy to World.
This allows game programmers to implement your own way of raycasting and AABB Querying aswell as how collision detection is handled.